PropertyBase as seperate file
and fix a bug in unification of D_spectrum/D_array data->device
